Use intrinsic function SUM for total_power
I was using in incremental summing within the loop of findpower to total up the total_power in watts with each iteration of the rectangle power 
calculation. I realized that I could use the intrinsic function SUM() on the resulting data array, and take out a line within the loop, and thus likely gain some efficiency in the long run. Learning to use these intrinsic array functions will likely be a benefit in the long run, for vectorizing operations.