Checked estimate, areas good
Area computation with rectangles is correct. Area computed is reported to user and written to output file. The actual area (pi*radius^2) is also reported for comparison and reported to user at the terminal and in the output file. 

A check of the "reasonableness" of the output power calculation was made against a real world case. A turbine produced by Vestas was examined and a good result for the estimate was output by the code that compares well with the estimate from the manufacturer.